backend/requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
djangorestframework==3.10.3
idna==2.8
pytz==2019.2
requests==2.21.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
urllib3==1.24.1
